fix: Fix compilation errors in cam-isp
- Fix libc::ioctl argument type in cma_buffer.rs: change ION_IOC_ALLOC_CMD
  and ION_IOC_SHARE_CMD from as u64 to as i32 (matches libc Ioctl type
  on Linux)
- Fix warp_engine.rs: use platform-independent 0 as libc::c_char instead
  of 0i8 for error_msg initialization

+++ cam-rust/cam-isp/src/mnn_buffer/cma_buffer.rs
@@ -314,7 +314,7 @@ impl CMABufferInfo {
         let fd = unsafe {
             libc::ioctl(
                 ion_fd.as_raw_fd(),
-                ION_IOC_ALLOC_CMD as u64,
+                ION_IOC_ALLOC_CMD as i32,
                 &alloc_data as *const _ as *mut libc::c_void,
             ) as RawFd
         };
@@ -366,7 +366,7 @@ impl CMABufferInfo {
         let share_fd = unsafe {
             libc::ioctl(
                 ion_fd,
-                ION_IOC_SHARE_CMD as u64,
+                ION_IOC_SHARE_CMD as i32,
                 &ion_fd as *const _ as *mut libc::c_void,
             ) as RawFd
         };
